Move vote_transaction out of src/
