+++ src/storage_stage.rs
@@ -279,7 +279,7 @@ mod tests {
     use storage_stage::NUM_IDENTITIES;
     use storage_stage::{get_identity_index_from_pubkey, StorageStage};
     use solana_sdk::vote_program::Vote;
-    use vote_transaction::VoteTransaction;
+    use solana_sdk::vote_transaction::VoteTransaction;
 
     #[test]
     fn test_storage_stage_none_ledger() {
